Use 64-bit time instead of 32-bit time for work queue utility.
Current timer has been upgraded to 64-bit but work queue utility
still uses 32-bit.
So timer->hitime can't be added rightly because the 64-bit time
is truncated by work queue utility.
"rmbox" will still receive 32-bit timestamp value only and can't
be upgraded to 64-bit timestamp.

+++ src/audio/dma-trace.c
@@ -97,7 +97,7 @@ static void trace_send(struct dma_trace_data *d)
 	trace_buffer("dts");
 }
 
-static uint32_t trace_work(void *data, uint32_t delay)
+static uint64_t trace_work(void *data, uint64_t delay)
 {
 	struct dma_trace_data *d = (struct dma_trace_data *)data;
