+++ src/include/reef/clock.h
@@ -53,7 +53,7 @@ uint32_t clock_get_freq(int clock);
 
 uint64_t clock_us_to_ticks(int clock, uint64_t us);
 
-uint32_t clock_time_elapsed(int clock, uint32_t previous, uint32_t *current);
+uint64_t clock_time_elapsed(int clock, uint64_t previous, uint64_t *current);
 
 void clock_register_notifier(int clock, struct notifier *notifier);
